+++ src/main/java/com/example/library/controller/LibraryRestController.java
@@ -1,7 +1,6 @@
 package com.example.library.controller;
 
 import java.io.IOException;
-import java.security.Principal;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.HashMap;
@@ -13,26 +12,23 @@ import javax.validation.Valid;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
+import com.example.library.dto.AuthorDto;
 import com.example.library.exception.*;
 import com.example.library.model.Actuality;
 import com.example.library.repository.AuthorRepository;
 import com.example.library.repository.BookRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.library.dto.AuthorDto;
+import com.example.library.dto.AuthorFullDto;
 import com.example.library.dto.BookDto;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
@@ -53,13 +49,13 @@ public class LibraryRestController {
     private Integer ACCESS_TOKEN_TIME;
 
     @GetMapping
-    public List<AuthorDto> findAll() {
-        return authorRepository.findAllByActuality(Actuality.ACTIVE).stream().map(author -> modelMapper.map(author, AuthorDto.class)).toList();
+    public List<AuthorFullDto> findAll() {
+        return authorRepository.findAllByActuality(Actuality.ACTIVE).stream().map(author -> modelMapper.map(author, AuthorFullDto.class)).toList();
     }
 
 
     @PostMapping("/authors")
-    public AuthorDto addAuthor(@RequestBody @Valid AuthorDto dto) {
+    public AuthorDto addAuthor(@RequestBody @Valid AuthorFullDto dto) {
         if (authorRepository.findEqualsAuthors(dto.getEmail(), dto.getName(), dto.getSecondName(), 0L).isPresent()) {
             throw new AuthorNotUniqueException();
         }
